requirements.txt
ascii-magic
Pillow